Make systemd socket activation actually work
When trying socket activation for my Docker network plugin, I noticed
that the first request to the plugin never reached the plugin. As it
turns out, the relevant code for extracting the passed-in socket from
systemd never got called. I think this caused systemd to failover,
cede the socket and try running the plugin normally, but at that
point, docker already hangs.

Signed-off-by: Steffen Müthing <steffen.muething@iwr.uni-heidelberg.de>